move the add question button
