+++ client/src/components/pages/NewSurvey.js
@@ -4,50 +4,12 @@ import { Page } from "@shopify/polaris";
 import { SurveyForm } from "..";
 
 export function NewSurvey() {
-  const [questions, setQuestions] = useState([]);
-  const addQuestion = type => {
-    const values = [...questions];
-    values.push({ type });
-    setQuestions(values);
-  };
-  const getQuestions = () => {
-    return [...questions];
-  };
   return (
     <Page
       title="Create a new survey"
       breadcrumbs={[{ content: "Admin Dashboard", url: "/app/admin/surveys" }]}
-      actionGroups={[
-        {
-          title: "Add Question",
-          accessibilityLabel: "Add a question",
-          actions: [
-            {
-              content: "Range Question",
-              accessibilityLabel: "Add a range question",
-              onAction: () => {
-                addQuestion("range");
-              }
-            },
-            {
-              content: "Option Question",
-              accessibilityLabel: "Add an option question",
-              onAction: () => {
-                addQuestion("option");
-              }
-            },
-            {
-              content: "Text Question",
-              accessibilityLabel: "Add a text question",
-              onAction: () => {
-                addQuestion("text");
-              }
-            }
-          ]
-        }
-      ]}
     >
-      <SurveyForm setQuestions={setQuestions} getQuestions={getQuestions} />
+      <SurveyForm />
     </Page>
   );
 }
